requirements.txt
flask
python-dotenv
azure-cognitiveservices-vision-face
opencv-python
tensorflow
pillow
numpy


